Restore _em_admin category

+++ scripts/compress_pdbml_vrpt_worker.sh
@@ -63,7 +63,9 @@ do
    mkdir -p $div_dir
   fi
 
-  mv -f $xml_file $div_dir && gzip $div_dir/$pdb_id-validation-full.xml
+  if [ ! -s $xml_file ] ; then
+   mv -f $xml_file $div_dir && gzip $div_dir/$pdb_id-validation-full.xml
+  fi
 
  fi
