+++ scripts/transl_to_mmcif_vrpt_alt_worker.sh
@@ -76,10 +76,12 @@ do
     mkdir -p $div_dir
    fi
 
-   mv -f $WORK_DIR/$mmcif_vrpt_file $div_dir && gzip $mmcif_vrpt_div_file
+   if [ ! -s $WORK_DIR/$mmcif_vrpt_file ] ; then
+    mv -f $WORK_DIR/$mmcif_vrpt_file $div_dir && gzip $mmcif_vrpt_div_file
 
-   if [ $proc_id_mod = 0 ] ; then
-    echo -e -n "\rDone "$((proc_id + 1)) of $total ...
+    if [ $proc_id_mod = 0 ] ; then
+     echo -e -n "\rDone "$((proc_id + 1)) of $total ...
+    fi
    fi
 
   fi
