+++ scripts/transl_to_rdf_vrpt_alt_worker.sh
@@ -82,10 +82,12 @@ do
     mkdir -p $div_dir
    fi
 
-   mv -f $rdf_vrpt_file $div_dir && gzip $rdf_vrpt_div_file
+   if [ ! -s $rdf_vrpt_file ] ; then
+    mv -f $rdf_vrpt_file $div_dir && gzip $rdf_vrpt_div_file
 
-   if [ $proc_id_mod = 0 ] ; then
-    echo -e -n "\rDone "$((proc_id + 1)) of $total ...
+    if [ $proc_id_mod = 0 ] ; then
+     echo -e -n "\rDone "$((proc_id + 1)) of $total ...
+    fi
    fi
 
   fi
